pyproject.toml
[project]
name = "fairness-pipeline-toolkit"
version = "0.1.0"
description = "A comprehensive toolkit for implementing fairness-aware machine learning pipelines"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "fairlearn>=0.10.0",
    "mlflow>=2.0.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "jupyter>=1.0.0",
    "notebook>=6.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "ruff>=0.1.0",
    "mypy>=0.910"
]
notebook = [
    "jupyter>=1.0.0",
    "notebook>=6.4.0",
    "ipykernel>=6.0"
]

[project.scripts]
run-fairness-pipeline = "fairness_pipeline_toolkit.cli:main"

[tool.ruff]
line-length = 88
target-version = "py313"
